+++ test/jest.config.js
@@ -21,7 +21,7 @@ module.exports = {
   // collectCoverage: false,
 
   // An array of glob patterns indicating a set of files for which coverage information should be collected
-  // collectCoverageFrom: ['src/**/*.js'],
+  collectCoverageFrom: ['../src/**/*.js'],
 
   // The directory where Jest should output its coverage files
   // coverageDirectory: null,